Move the SSAFOptions.h from the SSAF Frontend to the clang Frontend component
This resolves the library layering violation.
